build.sh: Fix user-space install
For user-space installation we need to set DESTDIR
and MANDIR as well. Use system default but let the user
change these if it wants to.

Nit: Changed the flags a bit (share dir is via -S and sys
dir is via -D) to distinguish between user and kernel params.

Move the defaults settings to the top and not in read_args()

Also, fixup the usage routine a bit to print the defaults
and the ${IAM} correctly.

And print an error message in case of an unknown param.

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>
Signed-off-by: Max Gurtovoy <maxg@mellanox.com>