icl, iscsi: Don't queue queue logout if we are not connected
We don't want to queue logout (iscsi_session_logout) in case
we are not connected. Thus, check we are connected before we
send the logout.

Since this call is under the session lock (which is also the
connection lock), have icl_conn_connected require the caller
to lock (OK given we just added the only call for it).

Signed-off-by: Sagi Grimberg <sagig@mellanox.com>