iser: Count sg list size properly
Instead of prior calculation of sg nents we assign its value after
setting the entire buffer to sg list and mark the end of the sg list.
This is needed to avoid cases of wrong calculation of unaligned
buffers.
Bug was detected by running diskinfo utility that sent 124k buffer
that was spread on 32 pages (calculation using DIV_ROUND_UP showed 31).

Signed-off-by: Max Gurtovoy <maxg@mellanox.com>
Signed-off-by: Sagi Grimberg <sagig@mellanox.com>